pyproject.toml
[tool.poetry]
name = "totrip"
version = "0.1.0"
description = ""
authors = ["Айтур Максутов <aitur1311@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
psycopg2 = "^2.9.10"
pyjwt = "^2.9.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
